+++ apps/remix/app/root.tsx
@@ -26,6 +26,7 @@ import popoverStyles from '@launchpad-ui/popover/styles.css';
 import tooltipStyles from '@launchpad-ui/tooltip/styles.css';
 import notificationStyles from '@launchpad-ui/notification/styles/Notification.css';
 import bannerStyles from '@launchpad-ui/banner/styles.css';
+import clipboardStyles from '@launchpad-ui/clipboard/styles.css';
 
 export const links: LinksFunction = () => {
   return [
@@ -44,6 +45,7 @@ export const links: LinksFunction = () => {
     { rel: 'stylesheet', href: tooltipStyles },
     { rel: 'stylesheet', href: notificationStyles },
     { rel: 'stylesheet', href: bannerStyles },
+    { rel: 'stylesheet', href: clipboardStyles },
   ];
 };