+++ packages/alert/__tests__/Alert.spec.tsx
@@ -10,7 +10,7 @@ import { Alert, AlertKind, AlertSize } from '../src';
 const createComponent = (props?: AlertProps) => <Alert {...props}>Alert text!</Alert>;
 
 describe('Alert', () => {
-  it('should render an alert with text', () => {
+  it('renders an alert with text', () => {
     render(createComponent());
     expect(screen.getByText('Alert text!')).toBeInTheDocument();
 
@@ -20,13 +20,13 @@ describe('Alert', () => {
   });
 
   describe('with dismissible', () => {
-    it('should render close button with dismissible', () => {
+    it('renders close button with dismissible', () => {
       render(createComponent({ dismissible: true }));
       const closeButton = screen.getByLabelText('Close this alert.');
       expect(closeButton).toBeInTheDocument();
     });
 
-    it('close button should trigger onDismiss when clicked and hide Alert', async () => {
+    it('triggers onDismiss when close button is clicked and hides the alert', async () => {
       const onDismiss = vi.fn();
       render(createComponent({ dismissible: true, onDismiss }));
 
@@ -47,12 +47,12 @@ describe('Alert', () => {
       expect(results).toHaveNoViolations();
     });
 
-    it('should render with role status by default', () => {
+    it('renders with role status by default', () => {
       render(createComponent({ size: AlertSize.MEDIUM }));
       expect(screen.getByRole('status')).toBeInTheDocument();
     });
 
-    it('should render with role alert for the error variant', () => {
+    it('renders with role alert for the error variant', () => {
       render(createComponent({ kind: AlertKind.ERROR }));
       expect(screen.getByRole('alert')).toBeInTheDocument();
     });