+++ packages/tooltip/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @launchpad-ui/tooltip
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`3ebbc38`](https://github.com/launchdarkly/launchpad-ui/commit/3ebbc38cb1dab6c1c3da1f6285fb7056a9be1d4d)]:
+  - @launchpad-ui/popover@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes